api/auth: integrated with gitosis package
When adding or removing a key, it will be added/removed from the gitosis
configuration file.

When running all tests together, there is still a little issue: since
the tests run in paralell, there is a race when accessing gitosis
repositories. I'm not sure yet about the best way to solve this race.