gitosis: starting to implement the gitosis.conf change agent
This agent will receive changes in the "Changes" channel and proccess
them. The change is a simple struct with three fields: Kind, Args and
Response.

Kind is an integer, but you should not use a magic number, but one of
the package constants for messages (AddKey, RemoveKey, etc.).

Args is a map of string to string, containing the arguments for the
given change. For instance, the AddKey change, requires the arguments
"key" and "member".

Response is a string channel. If the change returns anything, it will be
sent thought this channel. The AddKey change returns the name of the
file where the key was stored, and the caller should provide a channel
to receive this.

I've also unexported the functions deleteKeyFile and buildAndStoreKeyFile.
These functions are not supposed to be accessed directly anymore. The
same will happen to other functions in this package.