gitosis: rewrite agent as a type, and using interface for isolation
The Agent now is a type that contains a manager, which is an interface
providing methods for repository management. To provide a new repositoy
"backend", all you need to do is to provide another implementation for
the interface.

The code includes gitosisManager, the gitosis implementation for the
manager.

It closes issue #8 isolating agent tests from the gitosis.conf file (the
third option enumerated in the issue). The recordingManager just records
all actions, and the tests of the agent check only if the manager was
called with the right method. gitosisManager has its tests to guarantee
that it works, writing the gitosis.conf file.

We still need some improvements. The gitosisManager now includes a mutex
that will be used to synchronize access to the gitosis repository.