api/auth: added benchmark trying to reproduce issue #9
I was not able to reproduce the problem. It may be gone since the agent
got almost rewritten sooner today. Anyway, I'm not closing the issue
right now, lets give it a time.

We should add it to the continuous integration server, so if the bug
regret in the future, we will know :)

Update issue #9.