api/app: added SetEnv handler
This handler saves one or more environment variable to the database,
associating these with the app.

It reads all variables from the body of the request, in the following
format:

    NAME=value_of_variable

If the body of the request is empty, it returns 400 (bad request). It
also validates if the app exists (returning 404 - not found - if not),
and if the user has access to the app (returning 403 - forbidden - if
not). If any unknown error happen, it returns 500 (internal server
error).

At this moment, the handler does not write the environment variable(s)
in the unit. This will be a future step.

Update issue #35.