api/app: sending the environment variable to the environment agent
When setting an environment variable via handler, the handler sends the
change to a channel that runs a command to update the environment file
in the unit of the app.

In the future, when we support more than one unit per app, we will need
to run the command in all units of the app.