cmd: fix GetPassword
Now it is portable: in Go 1, Linux defines the Termios struct in the
syscall package, while all other platforms does not. So I need to
declare it by hand.

For that, I created a new term package inside the cmd package, and
declared a Termios struct in that package. The Termios might be a type
based on syscall.Termios (on Linux), or a brand new struct (on darwin,
freebsd, openbsd and netbsd).

It fixes issue #37, because now getPassword works on Lion. I'm not
really sure why.

Closes #37.