api/app: "unembedding" teams from app
The teams slice in the App type now is a slice of strings, referencing
multiple Teams in the "teams" collection. This change impacts many
tests, because in their code no team was being saved in the database,
and now they have to.

I know this is a huge commit, and I'm sorry about this, but I do not
like to commit broken tests :)

Fixes #41.