api/service: saving user's team in the ServiceInstance
The list of teams that has access to a ServiceInstance is an
intersection of all teams that the user is member of and all teams that
has access to the service being instantiated.

For example, if there are six teams in the database: alpha, beta, delta,
gamma, pi and theta, and alpha, beta, delta and gamma has access to the
service mysql:

    service:
        - name: mysql
        - teams:
            - alpha
            - beta
            - delta
            - gamma

And the user fsouza@tsuru.com is member of the teams alpha, beta and
zetta:

    user:
        - email: fsouza@tsuru.com
        - teams:
            - alpha
            - beta
            - zetta

When fsouza@tsuru.com creates an instance of mysql, it will have all
teams that are in both sets (alpha and beta).