api/app: use createBucket and destroyBucket in createApp and app.destroy
I've also changed destroy to work when the app does not have any unit
(in case of failures, or incosistency, for example).

There is a lot of changes in tests, now the tests are more higienic,
each of them delete destroy all apps that it creates, and createApp and
app.destroy are not used in all tests anymore. If the test need only the
app in the database, then it can be inserted using db.Session.Apps().

It should close issue #69, but it's not merged in master yet. Once our
patchs to goamz get merged, we can merge this branch.