netqueue: added QueueServer
QueueServer is the server that will be used to handle messages sent to
the queue.

To start a new server, the developer should use StartServer. StartServer
creates a new server instance, start the listener and runs a loop in a
new go routine. This loop will use read function to handle new client
connections.

For retrieving messages, the developer should use the Message method. It
returns the message or the error. It accepts a timeout, representing how
much time the method will wait for the message to arrive.

The server also provides the method Close, that is used to close the
server, and Addr, used to get the Addr of the server (it's useful when
you use random ports).