api/app: added enqueueApprcRegeneration method
This method is used for enqueue apprc regeneration requests in the
queue. Tsuru enqueues it, and collector runs the queue. Actually, the
queue could run itself in tsuru, or even in a separate process, but
right now it's coupled to the collector.

This change is supposed to close #203, but the issue need some love. I'm
going to test it before closing the issue.