app: added rollbackItself method to action interface
This method indicates wheter backward of the current action should be
called if it fails (which is not a rollback, actually).

I'm fixing the not-reported issue regarding the deletion of apps: if an
app "gopher" exists and I try to create it, insertApp.forward will fail,
and its backward will be called, deleting the original "gopher" app
(!!!). I don't think that it's a good idea to allow users to delete each
other app...