api: fix bug in revoke access from team
When revoking access from a team, it should give gandalf only a list of
users that should really lose access to the git repository (a user that
is member of two teams that have access to an app will still have access
to the repository if one of those teams is revoked).

Fixes #259.