provision/juju: added first implementation to CreateELB
It still doesn't support VPC, but it already works with EC2 and
availability zones. Everything is defined in the config file (not sure
if this is the best approach).

We need to extend config package to support slices. At least slice of
strings.

Juju provisioner will store all load balancers in tsuru's database. It
might be a good idea in the future that we abstract the database,
providing some utility functions in the db package, so developers
writing a new provisioner that needs to deal with the database don't
need to dive into MongoDB or mgo details.

Related to #244.