provision/juju: create and destroy elastic load balancers on demmand
Removing units is going to be as easy as these two, but adding units is
a little more complex, it will involve interaction with the queue,
because when we add units to Juju, them don't get an InstanceId
immediately, so we need to wait for the next collector execution, then
we will be able to register the unit.

Related to #244.