provision: simplify provisioning interfaces
LBManager is not really necessary at this moment. I know we will
probably need this in future, but right now we don't need it! Load
balancing is a detail fully implemented in the provisioner.

I've also added Addr() method to the Provisioner interface, this will be
used by collector to get the IP of the app.