queue: fix bug that prevent Handler from stoping
In the old code, first the handler asked for messages in the queue, and
if it failed to get a message from the queue, it checked if the handler
state is "running". That means that even when we call handler.Stop(), it
would stop only there was no message left in the queue.

Now it checks the state of the handler before trying to reserve more
jobs from the queue.