provision/juju: use the queue to register new units in the load balancer
Whenever a new unit is added to the app, JujuProvisioner will put a new
message in the queue, containing the name of the app and the name of the
new units. When handling this message, the handler will use juju to get
the Amazon instance-id of the units and register them in the load
balancer attached to the app.

Related to #247 (almost there, now we need to add new units in the load
balancer when we create a new app).