provision/juju: handle units that still don't have an ID
Whenever a unit doesn't have an instance id yet, the handler will add it
to a list of units that don't have ID. If none of the units have an ID,
the message is released, otherwise, all units that have an ID are
registered in the load balancer, the original message is deleted and a
new message is generated in the queue, for later processing.

Related to #247.