action: first version of the package
This version should be usable, but it is not used by app nor juju yet.
app and provision/juju packages should be the first client of this
package. It's a generalization of the actions "framework" provided
within the app package (not exported).

It includes some documentation that I believe is good for now.