app: starting to port to the new action framework
The first action ported is insertApp. I know this code looks like
JavaScript, but we might change it in the future. The renaming approach
is also odd, but I don't want to break the build, neither be unable to
run tests because of broken compilation.

Once all actions get ported, I will remove action.go, action_test.go,
actions.go and actions_test.go, and rename nactions.go and
nactions_test.go to actions.go and actions_test.go, respectively.

It will be a time of bad code, but it's a good reason :-)