app: remove dead code
This code looks stupid (and it is), but it's a legacy from the time that
we had two EnvVar types: one declared in the bind package, and other in
the app. So, SetEnvs existed to satisfy the interface bind.App, and used
SetEnvsToApp that received app.EnvVar, that's why the conversion
existed.

When we changed the implementation to use only bind.EnvVar, we probably
run some kind of sed or gofmt -r, and the lack of review allowed such
stupid code to exist.

Well, not anymore :-)