provision/juju: store units in the database
Using MongoDB for this is just stupid. We may use a key value storage in
the future.

This is the first step in dealing with Juju self-healing capabilities.
Whenever a instances is deleted or stopped, Juju creates another, with
the same name (and in the same security group). We need to detect this
change, remove the old instance from the load balancer and add the new
instance to it.