provision/juju: "healing" units as they change
When we stop or terminate an instance in Amazon, Juju spawns another
unit to replace the dead machine. A new instance have a different ID,
but the same unit name. We use a change in ID to detect that we need to
remove the old instance from the load balancer, add the new to the load
balancer and enqueue regeneration of apprc and start of the application
server.