provision/juju: set app-repo option in Provision instead of AddUnits
Whenever juju restores an unhealthy machine, tsuru detects it and make
sure that the code will be available in the new machine. We use the
charm option "app-repo" to indicate from which git repository the new
machine will clone code (during install hook).

The problem in the previous code is that it set this option during
unit-add operation, that means that an app that has always have only one
unit, it would never have the option "app-repo" defined. The solution is
simple: just set the option during the provisioning of the app.

Fix #326.