provision/docker: updating information of containers on CollectStatus
When we restart Docker daemon, it allocates a new IP and a new host port
for all running containers. Whenever the port to a container changes,
hipache is not able to access it anymore. We need to remove the old
route and readd it.

Updating the IP is also important, because we try to connect at the
allocated port in CollectStatus (net.Dial). Updating the IP in the
database prevents the unit from become "installing" (after a collector
execution).

Last, but not least, in order to detect changes in the host port, we
need to store it in the database (I've committed it in
88eda7f13079369ef0136ad4e99b4f41c268c2c7).