docs/faq: some formatting and text changes
All paragraphs are now at most 80-column wide. I've also removed some
text. Some answers contained implementation details (like the queue in
environment variables section). This is not relevant in the FAQ, and we
should not have to change the FAQ when we change the implementation
detail.

There was also other stuff, like saying that the default
application-per-user quota value is 2. The default value for any quota
is "unlimited".