provision/docker: use time.Sleep as temporary solution for a race
Unfortunately I can't fix it right now. dockerProvisioner may try to
serialize the environment variables before sshd is up, 5 seconds should
be enough.

The ideal solution is to come back to the queue, but the package app
should ask the provisioner whether the application is available or not.
That should not be hard to change, but I can't fix it _now_ (and we need
it working).