db, log: create one collection for each app log
Our previous solution having every log from every app in one single
collection had some issues. We had indexes too large and the ordering
wasn't preserved because we didn't have enough resolution on the time
field. Also setting a limit to the number of log entries for each app
was problematic.

One possible solution was putting every log entry from an app in a
single document inside the logs collection, this could help with a limit
of entries for each app, but it wouldn't help with the ordering problem
and it would make filtering in the database impossible, filters would
have to be made in go code.

The solution we've chosen creates a capped collection for each app with
the "logs_" prefix. This solves the cap and the ordering problems, since
insertion order is preserved in capped collections. It also improves
insert time since we have fewer indexes. The drawback is that running
`show collection` in Mongodb can get confusing.