Allow using = in the end of an env var's value
Environment variables that contain an equal sign (=) in the end are
now matched. For example:

BASE64_STRING=t5urur0ck5==

This change breaks the usage of env-set. If a given value contains any
space, it *has* to be enclosed by either single or double quotes.

This is ensured by asserting the equality of both the number of arguments
provided by the shell (i.e. bash) and the number of matches found.

Test and usage example were updated accordingly.