api: add sentinel to the deploy endpoint
It's needed so we can detect whether the deploy failed or succeeded. We
cannot rely on the HTTP status code, because it's sent before and then
the server starts to stream the response. The next step is to adapt the
pre-receive hooks to fail when the deploy is not successfully executed.
Something like:

	% curl tsuru-deploy-url | tail -1 | grep -q "^OK$"

Related to #442 and #701.