provision: add method SetUnitStatus, to redefine the status of a unit
We will use this method in a handler to set the status of a unit. This
handler will be called from a circus plugin.

The plugin will always send the status, like a heart beat. Then we will
be able to implement some logic based on this behaviour. For example: if
circus calls the API every one minute, we can check units that the last
call is from 3 minutes and set them automatically to error.

Related to #854.