provision: consider StatusError as available
This may see strange, but the way we handle "error" today requires this
approach: a unit changes to the error state whenever circus reports that
the application process is not running, meaning that the application is
down, but the unit is not. Hence, the unit is available, and we can run
commands against it.

Closes #876.