makes unbind match bind api signature call
The old way wasn't consistent, bind passed the app hostname and unit
host in the request body, while unbind passed only the unit host
in the querystring. Now both pass their parameters in the request body.

Change in the bind.Binder interface were needed, so we could receive the
app when unbinding an unit.

Relates to #794