Add ability to guess app name from dirname
This is *in addition* to guessing app name from the `tsuru` git remote.

The `tsuru` git remote is tried first and if that doesn't work, then we
take the name of the current directory and look to see if there is an
app with that name. If there is, we use that; else error out.

The main reason for adding this is that I am thinking of adding another
command that would add the git remote for the user.