Log to stderr
in addition to syslog or file

Some folks don't use syslog. They may want to run `tsr` using a process
manager like supervisord that manages log files for output to
stdout/stderr.

Also it is much easier to look at the console when running `tsr api` in
the terminal when developing or debugging.

E.g.:

    $ ~/go/bin/tsr api
    ...
    2014-12-27T07:13:12.851705809-08:00 GET /auth/scheme 200 in 0.519067ms
    2014/12/27 07:13:16 DEBUG: Ignored invalid token for /users: Invalid token
    2014/12/27 07:13:16 DEBUG: Failed to create user in the git server: Failed to connect to Gandalf server, it's probably down.