`tsr api` fail with error if config file not found
`tsr help` and `tsr token` will not fail if a config file is not found,
but `tsr api` will fail immediately with an error that if the config file is
not found. The error message will tell the user where tsuru looked for
the config file that was not found.

Before:

    [marca@marca-mac2 tsuru]$ ~/go/bin/tsr api
    Error: Config Error: you should have "listen" key set in your config file

After:

    [marca@marca-mac2 tsuru]$ ~/go/bin/tsr api
    2015/01/02 11:24:13 Opening config file: /etc/tsuru/tsuru.conf
    2015/01/02 11:24:13 Could not open tsuru config file at /etc/tsuru/tsuru.conf (open /etc/tsuru/tsuru.conf: no such file or directory).
      For an example, see: tsuru/etc/tsuru.conf
      Note that you can specify a different config file with the --config option -- e.g.: --config=./etc/tsuru.conf

Note that this message tells the user the real problem - that the config
file was not present, and it tells the user where it looked. Other
commands are not affected:

    [marca@marca-mac2 tsuru]$ ~/go/bin/tsr help
    tsr version 0.9.1.

    Usage: tsr command [args]

    Available commands:
      api
      help
      token
      version

    Use tsr help <commandname> to get more information about a command.

    [marca@marca-mac2 tsuru]$ ~/go/bin/tsr token
    252eb6231376253f50c82074061ad169f463ffba