provision/docker: add healthcheck for Docker
We will check only if there's at least one node in the cluster, and if
there's one node, we will ping it.

If the user have more than one node, he/she should use something like
`tsuru-admin docker-node-list` to check the status of the nodes, not the
healthcheck.

Related to #967.