repository: starting the RepositoryManager
This will be better than spreading ifs around, and will allow us to
isolate code related to Gandalf in a repository/gandalf package. Next
step is to implement the GandalfManager, and then make the packages use
repository.Manager instead of directly calling Gandalf.

This change also includes the nopManager, a manager that doesn't do
anything, and is part of the solution for issue #883.

The value of the "repo-manager" setting will define the behavior of the
Manager function, that will be:

  * when "repo-manager" is not defined, tsuru will use Gandalf. This is
    a guarantee of backward compatibility, allowing users to upgrade
    tsuru without having to change the configuration file
  * when "repo-manager" is defined to an unknown value (something that
    is not "gandalf" neither "nop"), it will use the nopManager and run
    without any actual repository management
  * when "repo-manager" is defined to a registered manager, just use
    requested manager