provision/docker: fix resize
StartExec blocks, so we need to call it inside a goroutine and then call
resize. Resize cannot be called if exec is not running, so we need to
start the goroutine with the StartExec call, wait the exec status to
change to Running and then resize TTY. It can still fail, but at least
it now works for longer sessions.