provision/docker: eliminate global mutexes in provisioner
Now provisioner and cluster initialization code is called by
Provisioner.Initialize(). Previously both container cluster and db
collection initialization were lazy, requiring a sync.Mutex to ensure it
only happened once. The downside is that tests are a bit more verbose,
as they need to ensure the provisioner instance is initialized.

Also, the global Mutex used for application lock in container movements
was changed to belong to the current movement request.