api: add an index page
This page will display some information to users willing to use the API.
This is useful so a user can open the tsuru API in the browser and have
a clue on how to use it. It's better than a 404 error, I think.

It is possible to customize the template that will be used, and it's
also possible to disable this feature at all, by defining the
configuration parameter "disable-index-page" as true.