cmd: flush pagerWriter if stdin is read, fixing confirmation prompts
Also, TSURU_PAGER is only started if the result is more than one page
long.

This was changed because the default pager, `less -RFX`, always added a
newline to the output even if the current result is less than one page
long. I didn't find a way to change this behavior and this made prompts
awkward as the cursor would always be on the next line, e.g.:

	$ tsuru login
	Email:
	admin@example.com