provision/docker: limit the amount of goroutines used for containers
When using numbers like 1000 containers, we do have some issues with it.
We're now introducing a setting for customizing the maximum amount of
woekrs. This isn't a pure "worker" architecture, as some workers may
work harder than others, because we just split the work across then
equally.

Related to #1149. unit-add doesn't use this code, and I also need to
document it.