Do not state number of steps
This was out of synch, as it said there are two steps but there are now in fact three. However, @fsouza has pointed out that we do not need to state the number of steps at all, and this change will mean no extra updates need to be made if a bullet is added or removed.