rec/rectest: use SetUpTest and TearDownTest to ensure values in db
While we do use different databases for each test, the configuration is
still global and shared across all tests, which is still something bad
when we run go test ./... I wonder if we can come up with a better
solution for this, like having per package config.

I also don't wanna have a big change in design just to make tests
happier. config is meant to be global and read-only, and the tests
should play nice with that.