provision/docker: partial support for the new Docker registry
We now support Docker distribution for deployment and platform
creation/update. We still need to port the image remove feature. This
version of the code already allows people to use tsuru with Docker 1.6
and Registry API 2.0, but won't remove any images from the Registry
(when the app is removed).

Related to #1175.