api: Fix intermittent SwapApp test failure
`TestSwapApp1Locked` has been intermittently failing for me. I can't
reproduce it when running the test by itself, but can when run in
conjunction with `TestSwap`:

    ➜  tsuru git:(master) go test github.com/tsuru/tsuru/api -check.f TestSwap -check.v
    PASS: app_test.go:2853: S.TestSwap      0.486s

    ----------------------------------------------------------------------
    FAIL: app_test.go:2877: S.TestSwapApp1Locked

    app_test.go:2895:
        c.Assert(err, check.ErrorMatches, "app1: App locked by x, running /test. Acquired in .*")
    ... value = nil
    ... regex string = "app1: App locked by x, running /test. Acquired in .*"
    ... Error value is nil

    ----------------------------------------------------------------------
    PASS: app_test.go:2898: S.TestSwapApp2Locked    10.097s
    PASS: app_test.go:2962: S.TestSwapIncompatibleAppsForceSwap     0.034s
    PASS: app_test.go:2915: S.TestSwapIncompatiblePlatforms 0.041s
    PASS: app_test.go:2938: S.TestSwapIncompatibleUnits     0.032s
    OOPS: 5 passed, 1 FAILED
    --- FAIL: Test (10.98s)
    FAIL
    FAIL    github.com/tsuru/tsuru/api      10.989s

I'm thinking that this is a race condition to do with deleting the
(unlocked) app from the previous test and inserting the new (locked) app
directly into the database. Or `app.Provisioner.Provision()` introduced in
6fe0fc6 is unlocking the app afterwards?

Either way, this seems to be fixed by using `app.CreateApp()` and
`deleteApp()` instead of manipulating the database directly. Which matches
with what is done in `TestSwap`.