provision/docker: fix the usePlatformImage function
The container presence checking is not used anymore, as we keep an
history of images in the database. This fixes a bug that caused tsuru to
always use the platform image when the application has no units. It's
the right behavior after the app has been created, but it's wrong when
user removes all application units.

This behavior had no effect before tsuru-server 0.11, but since this
release, users are now able to remove all units from an application.