provision/docker: trust Docker to check if the container is running
Instead of checking for the container process, we're just checking its
state. Since the container runs just the process that it's required to
run, we don't need to check for the process presence in the top output.
The code now just checks whether the container is running or restarting
in Docker.

Related to #647 and #1136.