router/vulcand: Stubs for interface and tests
Very first step in implementing vulcand as a router. This is desirable
because it implements TLS and SNI; support for which will be added at a much
later date.

Satisfies all methods of the `router.Router` interface but doesn't actually
do anything. The empty tests are based on these interface methods and don't
necessarily reflect all of the scenarios that we need to test, ie. address
versus URL parsing for `AddRoute()`.