router/vulcand: Setup in-memory vulcand for tests
This runs a vulcand server with an in-memory datastore (instead of etcd)
that we can inspect during tests, eg. `s.engine.GetHosts()`.

Initialised for each test in `SetUpTest` because it ensures that there is no
state leaked between tests, which may make our tests order dependent and
introduce side effects.