router/vulcand: Cleanup on partial AddBackend
Adding a backend is a two phase process because we need to add a "frontend"
and then a "backend". If the second of these operations fails then we should
cleanup the first operation.

The test simulates this by making the API respond with an HTTP 500 error to
the second `POST` request and then checks that there are no frontends or
backends. This makes the test agnostic of which way round the operations
have been implemented.