provision/docker: fix bug in the calculation of nodes gap
The intermitent test TestAutoScaleConfigRunRebalanceOnly actually
exposed a bug: using 0 as the sentinel for the minimum value isn't good,
because it's a valid value for the amount of containers in a node.

In the test, there are two node: one with 4 containers and other with 0
containers, which means that the gap is 4. The bug is: when the host
with 0 containers is checked in the first iteration, the gap ends being
0. The test is intermitent because we use a map to store the containers,
so there were two scenarios:

 1. check the host with 4 containers, then the host with 0 containers,
 the test passes
 2. check the host with 0 containers, then the host with 4 containers,
 the test fails

I preferred to keep the sentinel logic and use -1 instead of just using
math.MaxInt64 with no sentinel.  No special reason, just safer (I don't
expect to see math.MaxInt64 containers in a single node, but let's take
the safe path anyway).