provision/docker: safer remove units implementation
Previously, if RemoveUnits failed while removing routes from router the
removal process would continue ignoring errors. This would cause the
router to keep sending requests to units that were destroyed.

Now we only destroy units from docker if the process of removing the
routes was successful, using the same pipeline actions already used
during deploy.

If we fail to remove some routes we try to add back the already removed
ones and the pipeline is aborted. This makes sure that all units
referenced in the router will still exist.

However, if the process of adding removed routes back to the router
after a failure also fails, we may have a router that will not be using
all possible units, this could cause a performance problems but the
application won't fail. To mitigate this problem one possibility is as
periodic process to refresh routers based on units in the database. We
should explore this idea in the future.