provision/docker: tsuru allocating container ports instead of docker
For now, to enable this feature it's necessary to set the config
`docker:port-allocator` to `tsuru`.

The motivation behind this feature is making sure the host port mapped
to one container never changes, even after restarting docker daemon or
rebooting the host. This way, we can always be sure that routers are
pointing to a valid address.

Usually this isn't a big problem since tsuru server will receive a ping
from the container and fix its mapped host port on routers if it has
changed. However, if the tsuru api server is down there's a big chance
of routers having incorrect entries.

With this change tsuru api server is not a SPOF if some catastrophic
failure happens (e.g. datacenter power failure).

Related to #1072.