provision/docker: change scheduler behavior if max memory is reached
Fixes #1251.

Now memory aware scheduler will behave differently if node auto scale is
enabled. If it's enabled the scheduler will not fail to schedule a node
if it has reached the max memory. In this case, going over max will only
happen temporarily. The auto scale process will start and a new node
will be created to accommodate the extra memory requirement shortly
after.

If node auto scale is disable though, tsuru will enforce the max memory
policy because this will trigger an error and someone will have to
manually add a new node to allow new units to be created.