provision/docker: remove multiple nodes on memory auto scale down
This also changes the way auto scale down works. Previously we would run
a simulation removing each node, rebalancing the containers and then
checking if they were inside our memory limits.

This was slow and didn't scale very well. If we wanted to check whether
multiple nodes could be removed, we would have to run the simulation
again after removing 1 node, and then again after removing 2 nodes and
so on. This would lead to an O(n!) execution time where n is the number
of nodes, thus not feasible.

Now we simply check how many nodes are necessary to fit the total
reserved memory plus one container using the largest plan. This is not
optimal but it's much simpler to compute.