provision/docker: move the container type to its own package
This is an effort to make the provision/docker package smaller, so we
can get faster tests and less code inside a package. I know that the
import path gets too big, but we might change it in the future.

Other components that we're going to move:

 * auto scale
 * healing (container and node)

Other things may move as well.