provision/docker: move bs related functions to a dedicated package
Moving commands and management functions. The handlers will remain in
the provision/docker package, as we want to move them later to a package
dedicated to handlers.

This is part of an effort to make the package provision/docker smaller.
See 43c5f2776488f119b1870fb05747584775aea844 for more details.