api, app: move granting/revoking logic away from the handler
Issue #1270 describes a bug with the SetTeamOwner method, but the issue
was deeper: SetTeamOwner already calls Grant, but Grant didn't behave as
expected. It's all due to some bad coding: app.Grant only changed the
app object, and relied on the handler for database updating and
repository integration, now everything is in the app package.

Fix #1270.