api: use the KeepAliveWriter in all of the streaming handlers
Properly handling smaller timeouts (we're used to 300 seconds timeouts -
which is definitely not good, I know, but there are some tsuru users
with 60 seconds timeouts, and this has been a problem in docker pull
after upgrading to Docker 1.7).

Maybe we should be able to handle timeouts smaller than 30 seconds?
(Probably making the interval in KeepAliveWriter configurable)