permission: add groundwork for new tsuru authorization system
Please be warned that this system is subject to lots of changes. The
code included in this commit is not complete and has the purpose of
helping us visualize how the authorization system will be used in tsuru.

The following concepts were introduced (again, names are subject to
change, naming things is hard):

PermissionScheme
================

It's the definition of a permission that can be used in tsuru. A
PermissionScheme consists of a name, a parent and a list of ContextTypes
that are valid for this PermissionScheme. PermissionSchemes can be
represented with a dot notation to show all its parents. e.g:
'app.update.env.set' represents a PermissionScheme named 'set' with a
parent named 'env' which has a parent named 'update' and so on.

Granting permissions for a PermissionScheme automatically grants
permission for all its children.

It will not be allowed for the end-user to register new
PermissionSchemes, tsuru will already ship with a predefined set of
PermissionSchemes, registered using the PermissionRecord, which will be
verified on demand when executing operations on resources.

PermissionRecord
================

Used to simplify registering nested PermissionSchemes and serves as an
in-memory repository of available PermissionSchemes.

ContextType
===========

A predefined set of possible types of Context that can be used when
creating a Permission if allowed by the PermissionScheme. Current
options include 'App', 'Team', 'Pool' and Global.

Context
=======

A association between a ContextType and one specific value. It's used
when creating Permissions that will be associated with an existing
authenticated Token. e.g. Context type 'Team' with value 'myteam'.

The Global ContextType is special where a Context created with it does
not need a value. A permission created using this Context will always be
valid for the specified PermissionScheme.

Permission
==========

A association between a PermissionScheme and a Context. Each existing
authenticated Token will be able to have a list of Permissions
associated with it. One example of a valid Permission would be:
'app.update.env' for the ContextType 'Team' with the value 'myteamname'.

During the execution of an authorization aware operations tsuru will
call the permission's module `Check()` function with:

* The Token trying to execute the operation.
* The PermissionScheme required to perform the operation. e.g
  'app.update.env.set'.
* All possible Contexts present during the operation. e.g. [Team
  'myteam', App 'myapp', Pool 'mypool']

The `Check()` call will then iterate on all Permissions available for
the Token and will only succeed if a Permission exist where it's a
parent of the specified PermissionScheme and include a Context matching
at least one of the possible Contexts.

Purposely missing is the definition of Roles. It was simpler to
visualize the system at this time as a direct association between Tokens
and Permissions. Roles will be integrated in this system soon and will
basically provide a way of aggregating permissions.

Related to #1278.